requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
flake8==5.0.4
flake8-builtins==2.1.0
flake8-django==1.1.5
flake8-import-order==0.18.2
flake8-print==5.0.0
Markdown==3.4.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.7.1
sqlparse==0.4.3
